AI/requirements.txt
tensorflow==2.2.0
Pillow==7.2.0
numpy==1.19.1
matplotlib==3.3.2
rasterio==1.1.7
pandas==1.1.3
